+++ src/R-yyjson-serialize.h
@@ -71,6 +71,7 @@ typedef struct {
   unsigned int null;
   int digits;
   int digits_secs;
+  int signif;
   bool auto_unbox; 
   unsigned int name_repair;
   unsigned int str_specials;